pyproject.toml
[project]
name = "payment_system"
version = "0.1.0"
description = "Test task for interview"
authors = [
    { name = "Nikita Evstropov", email = "evstropov.n@gmail.com" }
]
readme = "README.md"
requires-python = "~3.9"
dependencies = [
    "django (>=4.2.17,<5.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
]
